MobileApp/allrequirements.txt
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
baidu-aip==4.16.13
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
exceptiongroup==1.2.0
fastapi==0.110.1
flet==0.22.0
flet-contrib==2024.3.6
flet-core==0.22.0
flet-runtime==0.22.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
packaging==23.2
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.31.0
rich==13.7.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
watchdog==4.0.0
watchfiles==0.21.0
Wave==0.0.2
websockets==12.0
